+++ app/src/main/java/kr/pe/ssun/cokedex/ui/detail/PokemonUiState.kt
@@ -5,15 +5,15 @@ import kr.pe.ssun.cokedex.model.PokemonDetail
 sealed interface PokemonUiState {
     data class Success(
         val pokemon: PokemonDetail,
-        val colorStart: Long,
-        val colorEnd: Long,
+        val colorStart: Int,
+        val colorEnd: Int,
     ) : PokemonUiState
     data class Loading(
         val id: Int,
         val name: String,
         val imageUrl: String,
-        val colorStart: Long,
-        val colorEnd: Long,
+        val colorStart: Int,
+        val colorEnd: Int,
     ) : PokemonUiState
     object Error : PokemonUiState
 }
